Creating disk separately using qemu-img create instead of getting kvm to create the image, in order to be able to better specify parameters to gain better performance. This means that we no longer need to ask for sparse/dedicated storage and only use sparse as the only reason to use dedicated was for the performance.  - Haven't checked if this could lead to accidental over-provisioning, in which case the admin needs to be the judge. The admin may desire over provisioning of disk after all.
We now write the command(s) to a file in /tmp and get the user to call that (displaying the command to run that script). This prevents the user from having to do a massive copy and paste job since it is small enough that they can just type the command.

# Current issues
There are no checks in place to see if the user specified a name that is already taken. This will result in the script executing fine, but when the user calls the generated script, that one will fail.